+++ src/id.rs
@@ -341,7 +341,7 @@ impl fmt::Display for ParseError {
 
 #[cfg(feature = "std")]
 #[cfg_attr(docsrs, doc(cfg(feature = "std")))]
-mod std_ext {
+mod with_std {
     use super::{ParseError, Scru128Id};
 
     impl TryFrom<String> for Scru128Id {
@@ -596,7 +596,7 @@ mod tests {
 
 #[cfg(feature = "serde")]
 #[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
-mod serde_support {
+mod with_serde {
     use super::{fmt, str, Scru128Id};
     use serde::{de, Deserializer, Serializer};