+++ src/lib.rs
@@ -37,11 +37,10 @@
 //!
 //! - `std` configures [`Scru128Generator`] with the system clock. Without `std`, this
 //!   crate provides basic SCRU128 primitives available under `no_std` environments.
-//! - `rand` enables a blanket implementation for [`rand::RngCore`] to use `rand` and
-//!   any other conforming random number generators with [`Scru128Generator`].
-//! - `default_rng` (implies `std` and `rand`) provides the default random number
-//!   generator for [`Scru128Generator`] and enables the [`Scru128Generator::new()`]
-//!   constructor.
+//! - `rand` enables an adapter for [`rand::RngCore`] to use `rand` and any other
+//!   conforming random number generators with [`Scru128Generator`].
+//! - `default_rng` (implies `std`) provides the default random number generator for
+//!   [`Scru128Generator`] and enables the [`Scru128Generator::new()`] constructor.
 //! - `global_gen` (implies `default_rng`) provides the process-wide default SCRU128
 //!   generator and enables the [`new()`] and [`new_string()`] functions.
 //!
@@ -61,7 +60,7 @@ pub use id::{ParseError, Scru128Id};
 
 pub mod generator;
 #[doc(hidden)]
-pub use generator as gen;
+pub use generator as r#gen;
 pub use generator::Scru128Generator;
 
 /// The maximum value of 48-bit `timestamp` field.
@@ -73,8 +72,7 @@ const MAX_COUNTER_HI: u32 = 0xff_ffff;
 /// The maximum value of 24-bit `counter_lo` field.
 const MAX_COUNTER_LO: u32 = 0xff_ffff;
 
-#[cfg(feature = "std")]
-#[cfg(test)]
+#[cfg(all(test, feature = "std"))]
 mod tests {
     use crate::{Scru128Generator, Scru128Id};